Annual expression implies the 1st of the year
For practical use, it makes more sense that the intent of
the yearly expressions are limited to the 1st of the year.

Should "every November" also imply the 1st of the month?
For now, it remains active for the entire month of November.
This can be worked around by parsing "every November 1st" for
the time being. Language is tough.